move user service into separate user folder
